requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.1.0
psycopg2-binary==2.9.3
sqlparse==0.4.2
